refactor(xapiri): extract dashboard_chrome.go + dashboard_overlays.go
Chrome: renderTabBar, renderSysWidget, fmtBytes, fmtRate, tabAtX,
renderBottomStrip, renderFooter moved to dashboard_chrome.go.
Token re-prompt overlay moved to dashboard_overlays.go per ADR 0014.
